Move helpers into src/App/Helpers, add SSRHelper
